verify session befor crud googl meet

+++ lib/utils.ts
@@ -717,7 +717,13 @@ export const updateMeetingLink = async ({
     eventId,
     startDate,
     endDate,
+    session,
+    signIn
 }: any) =>{
+    if (!session) {
+      await signIn("google");
+      return;
+    }
     await fetch("/api/update-meet", {
         method: "PUT",
         headers: { "Content-Type": "application/json" },
@@ -732,7 +738,13 @@ export const updateMeetingLink = async ({
 
 export const deleteMeetingLink = async ({
     eventId,
+    session,
+    signIn
 }: any) =>{
+    if (!session) {
+      await signIn("google");
+      return;
+    }
     await fetch("/api/delete-meet", {
         method: "DELETE",
         headers: { "Content-Type": "application/json" },
